requirements.txt
fastapi==0.61.0
uvicorn==0.11.8
python-decouple==3.3